E2/app/requirements.txt
Flask==2.3.*
python-dotenv==1.0.1
psycopg[binary,pool]==3.1.*
datetime==5.5
cython>=0.29.24
gunicorn==22.0.0